chore: move filter tests into filter mod
