+++ crates/core/src/operations/merge/mod.rs
@@ -29,6 +29,7 @@
 //! ````
 
 use std::collections::HashMap;
+use std::fmt::Debug;
 use std::sync::Arc;
 use std::time::Instant;
 
@@ -1375,7 +1376,6 @@ mod tests {
     use crate::kernel::StructField;
     use crate::operations::load_cdf::collect_batches;
     use crate::operations::merge::filter::generalize_filter;
-    use crate::operations::merge::try_construct_early_filter;
     use crate::operations::DeltaOps;
     use crate::protocol::*;
     use crate::writer::test_utils::datafusion::get_data;
@@ -1389,18 +1389,14 @@ mod tests {
     use arrow_schema::DataType as ArrowDataType;
     use arrow_schema::Field;
     use datafusion::assert_batches_sorted_eq;
-    use datafusion::datasource::provider_as_source;
     use datafusion::physical_plan::ExecutionPlan;
     use datafusion::prelude::*;
     use datafusion_common::Column;
-    use datafusion_common::ScalarValue;
     use datafusion_common::TableReference;
     use datafusion_expr::col;
     use datafusion_expr::expr::Placeholder;
     use datafusion_expr::lit;
     use datafusion_expr::Expr;
-    use datafusion_expr::LogicalPlanBuilder;
-    use datafusion_expr::Operator;
     use itertools::Itertools;
     use regex::Regex;
     use serde_json::json;
@@ -1409,7 +1405,7 @@ mod tests {
 
     use super::MergeMetrics;
 
-    async fn setup_table(partitions: Option<Vec<&str>>) -> DeltaTable {
+    pub(crate) async fn setup_table(partitions: Option<Vec<&str>>) -> DeltaTable {
         let table_schema = get_delta_schema();
 
         let table = DeltaOps::new_in_memory()
@@ -2612,256 +2608,6 @@ mod tests {
         .eq(col(Column::new(target.clone().into(), "id")));
 
         assert_eq!(generalized, expected_filter);
-    }
-
-    #[tokio::test]
-    async fn test_try_construct_early_filter_with_partitions_expands() {
-        let schema = get_arrow_schema(&None);
-        let table = setup_table(Some(vec!["id"])).await;
-
-        assert_eq!(table.version(), 0);
-        assert_eq!(table.get_files_count(), 0);
-
-        let ctx = SessionContext::new();
-        let batch = RecordBatch::try_new(
-            Arc::clone(&schema),
-            vec![
-                Arc::new(arrow::array::StringArray::from(vec!["B", "C", "X"])),
-                Arc::new(arrow::array::Int32Array::from(vec![10, 20, 30])),
-                Arc::new(arrow::array::StringArray::from(vec![
-                    "2021-02-02",
-                    "2023-07-04",
-                    "2023-07-04",
-                ])),
-            ],
-        )
-        .unwrap();
-        let source = ctx.read_batch(batch).unwrap();
-
-        let source_name = TableReference::parse_str("source");
-        let target_name = TableReference::parse_str("target");
-
-        let source = LogicalPlanBuilder::scan(
-            source_name.clone(),
-            provider_as_source(source.into_view()),
-            None,
-        )
-        .unwrap()
-        .build()
-        .unwrap();
-
-        let join_predicate = col(Column {
-            relation: Some(source_name.clone()),
-            name: "id".to_owned(),
-        })
-        .eq(col(Column {
-            relation: Some(target_name.clone()),
-            name: "id".to_owned(),
-        }));
-
-        let pred = try_construct_early_filter(
-            join_predicate,
-            table.snapshot().unwrap(),
-            &ctx.state(),
-            &source,
-            &source_name,
-            &target_name,
-        )
-        .await
-        .unwrap();
-
-        assert!(pred.is_some());
-
-        let split_pred = {
-            fn split(expr: Expr, parts: &mut Vec<(String, String)>) {
-                match expr {
-                    Expr::BinaryExpr(ex) if ex.op == Operator::Or => {
-                        split(*ex.left, parts);
-                        split(*ex.right, parts);
-                    }
-                    Expr::BinaryExpr(ex) if ex.op == Operator::Eq => {
-                        let col = match *ex.right {
-                            Expr::Column(col) => col.name,
-                            ex => panic!("expected column in pred, got {ex}!"),
-                        };
-
-                        let value = match *ex.left {
-                            Expr::Literal(ScalarValue::Utf8(Some(value))) => value,
-                            ex => panic!("expected value in predicate, got {ex}!"),
-                        };
-
-                        parts.push((col, value))
-                    }
-
-                    expr => panic!("expected either = or OR, got {expr}"),
-                }
-            }
-
-            let mut parts = vec![];
-            split(pred.unwrap(), &mut parts);
-            parts.sort();
-            parts
-        };
-
-        let expected_pred_parts = [
-            ("id".to_owned(), "B".to_owned()),
-            ("id".to_owned(), "C".to_owned()),
-            ("id".to_owned(), "X".to_owned()),
-        ];
-
-        assert_eq!(split_pred, expected_pred_parts);
-    }
-
-    #[tokio::test]
-    async fn test_try_construct_early_filter_with_range() {
-        let schema = get_arrow_schema(&None);
-        let table = setup_table(Some(vec!["modified"])).await;
-
-        assert_eq!(table.version(), 0);
-        assert_eq!(table.get_files_count(), 0);
-
-        let ctx = SessionContext::new();
-        let batch = RecordBatch::try_new(
-            Arc::clone(&schema),
-            vec![
-                Arc::new(arrow::array::StringArray::from(vec!["B", "C"])),
-                Arc::new(arrow::array::Int32Array::from(vec![10, 20])),
-                Arc::new(arrow::array::StringArray::from(vec![
-                    "2023-07-04",
-                    "2023-07-04",
-                ])),
-            ],
-        )
-        .unwrap();
-        let source = ctx.read_batch(batch).unwrap();
-
-        let source_name = TableReference::parse_str("source");
-        let target_name = TableReference::parse_str("target");
-
-        let source = LogicalPlanBuilder::scan(
-            source_name.clone(),
-            provider_as_source(source.into_view()),
-            None,
-        )
-        .unwrap()
-        .build()
-        .unwrap();
-
-        let join_predicate = col(Column {
-            relation: Some(source_name.clone()),
-            name: "id".to_owned(),
-        })
-        .eq(col(Column {
-            relation: Some(target_name.clone()),
-            name: "id".to_owned(),
-        }));
-
-        let pred = try_construct_early_filter(
-            join_predicate,
-            table.snapshot().unwrap(),
-            &ctx.state(),
-            &source,
-            &source_name,
-            &target_name,
-        )
-        .await
-        .unwrap();
-
-        assert!(pred.is_some());
-
-        let filter = col(Column {
-            relation: Some(target_name.clone()),
-            name: "id".to_owned(),
-        })
-        .between(
-            Expr::Literal(ScalarValue::Utf8(Some("B".to_string()))),
-            Expr::Literal(ScalarValue::Utf8(Some("C".to_string()))),
-        );
-        assert_eq!(pred.unwrap(), filter);
-    }
-
-    #[tokio::test]
-    async fn test_try_construct_early_filter_with_partition_and_range() {
-        let schema = get_arrow_schema(&None);
-        let table = setup_table(Some(vec!["modified"])).await;
-
-        assert_eq!(table.version(), 0);
-        assert_eq!(table.get_files_count(), 0);
-
-        let ctx = SessionContext::new();
-        let batch = RecordBatch::try_new(
-            Arc::clone(&schema),
-            vec![
-                Arc::new(arrow::array::StringArray::from(vec!["B", "C"])),
-                Arc::new(arrow::array::Int32Array::from(vec![10, 20])),
-                Arc::new(arrow::array::StringArray::from(vec![
-                    "2023-07-04",
-                    "2023-07-04",
-                ])),
-            ],
-        )
-        .unwrap();
-        let source = ctx.read_batch(batch).unwrap();
-
-        let source_name = TableReference::parse_str("source");
-        let target_name = TableReference::parse_str("target");
-
-        let source = LogicalPlanBuilder::scan(
-            source_name.clone(),
-            provider_as_source(source.into_view()),
-            None,
-        )
-        .unwrap()
-        .build()
-        .unwrap();
-
-        let join_predicate = col(Column {
-            relation: Some(source_name.clone()),
-            name: "id".to_owned(),
-        })
-        .eq(col(Column {
-            relation: Some(target_name.clone()),
-            name: "id".to_owned(),
-        }))
-        .and(
-            col(Column {
-                relation: Some(source_name.clone()),
-                name: "modified".to_owned(),
-            })
-            .eq(col(Column {
-                relation: Some(target_name.clone()),
-                name: "modified".to_owned(),
-            })),
-        );
-
-        let pred = try_construct_early_filter(
-            join_predicate,
-            table.snapshot().unwrap(),
-            &ctx.state(),
-            &source,
-            &source_name,
-            &target_name,
-        )
-        .await
-        .unwrap();
-
-        assert!(pred.is_some());
-
-        let filter = col(Column {
-            relation: Some(target_name.clone()),
-            name: "id".to_owned(),
-        })
-        .between(
-            Expr::Literal(ScalarValue::Utf8(Some("B".to_string()))),
-            Expr::Literal(ScalarValue::Utf8(Some("C".to_string()))),
-        )
-        .and(
-            Expr::Literal(ScalarValue::Utf8(Some("2023-07-04".to_string()))).eq(col(Column {
-                relation: Some(target_name.clone()),
-                name: "modified".to_owned(),
-            })),
-        );
-        assert_eq!(pred.unwrap(), filter);
     }
 
     #[tokio::test]
